Migrate script to TypeScript
