+++ packages/config/src/Errors.ts
@@ -9,3 +9,28 @@ export class ConfigError extends Error {
     this.name = this.constructor.name;
   }
 }
+
+export function errorFromJSON({ name, ...json }: any): Error {
+  let error: any;
+  if (name === 'TypeError') {
+    error = new TypeError(json.message);
+  } else {
+    error = new Error(json.message);
+  }
+  for (const key of Object.keys(json)) {
+    if (key in json) {
+      error[key] = json[key];
+    }
+  }
+  return error;
+}
+
+export function errorToJSON(error: any): any {
+  return {
+    ...error,
+    message: error.message,
+    code: error.code,
+    name: error.name,
+    stack: error.stack,
+  };
+}
