+++ packages/config/src/__tests__/getConfig-test.js
@@ -1,24 +1,6 @@
 import { join } from 'path';
 
-import { findAndEvalConfig, serializeAndEvaluate } from '../getConfig';
-
-describe('serializeAndEvaluate', () => {
-  it(`serializes item`, () => {
-    expect(
-      serializeAndEvaluate({
-        foo: 'bar',
-        boo: true,
-        inn: 200,
-        then: [true, { foo: 'bar' }],
-        alpha: () => ({ beta: ['val'] }),
-        last: {
-          bar: 'foo',
-          charlie: [2, 'delta'],
-        },
-      })
-    ).toMatchSnapshot();
-  });
-});
+import { findAndEvalConfig } from '../getConfig';
 
 describe('findAndEvalConfig', () => {
   it(`throws a useful error for JS configs with a syntax error`, () => {
